app/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.112.0
fastapi-pagination==0.12.11
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
numpy==1.26.2
pandas==2.1.3
pillow==10.2.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pymongo==4.5.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2023.3.post1
reportlab==4.1.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.37.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
uvicorn==0.23.2
scikit-learn==1.5.0
GitPython ==3.1.43